kk/requirements_min.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-Dance==7.1.0
Werkzeug==3.0.1
bcrypt==4.1.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Pillow==11.0.0
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
psycopg[binary]
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0
phonenumbers==8.13.26
python-jose[cryptography]==3.3.0
cryptography==45.0.7
requests==2.31.0
sentry-sdk
# Needed for Roboflow plate blurring
opencv-python-headless
numpy
